Update README.md (#93)
https://blog.crypto.com/the-crypto-org-chain-ecosystem-grants-programme/

3. Awesome Projects Rewards Programme

The Awesome Project Rewards Programme allows community members to share and showcase their Crypto.org Chain projects. Rewards will be granted to projects that greatly benefit the community.

To be considered for this programme, the project must:

Benefit a group of or all community members
Not promote particular validators
To get involved, please refer to this page and submit your contribution accordingly.